project_Doniraj/faza5/requirements.txt
Django
pillow
django-cleanup
django-allauth
django-htmx
django-bootstrap-v5
shortuuid
